% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_overview_map.R
\name{create_overview_map}
\alias{create_overview_map}
\title{Title}
\usage{
create_overview_map(
  sample_frame_for_dashboard,
  admin_one_sf,
  admin_two_sf,
  admin_three_sf
)
}
\arguments{
\item{sample_frame_for_dashboard}{Sample frame data create by \code{create_sample_frame_file}}

\item{admin_one_sf}{admin boundary one}

\item{admin_two_sf}{admin boundary two}

\item{admin_three_sf}{admin boundary three}
}
\description{
Title
}
