% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_frame.R
\name{create_sample_frame_file}
\alias{create_sample_frame_file}
\title{Create sample frame file}
\usage{
create_sample_frame_file(
  sample_frame,
  base_file,
  completed_survey = "Number_survey",
  pop_group = "Population.Group",
  admin_name = "Strata.name"
)
}
\arguments{
\item{sample_frame}{Sample frame. You can use \code{cleaningtools::review_sample_frame_with_dataset} to create the file.}

\item{completed_survey}{Column name for completed survey}

\item{pop_group}{Column name for population group}

\item{admin_name}{Column name for Adming name}
}
\value{
Creates a data file
}
\description{
Create sample frame file
}
