% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_base_file.R
\name{creating_base_file}
\alias{creating_base_file}
\title{Title}
\usage{
creating_base_file(
  validated_analysis,
  stat = "stat",
  pop_group = "pop_group",
  admin_name = "admin_name",
  admin_level = "admin_level",
  analysis_key,
  kobo_survey,
  kobo_choice,
  dap,
  dap_sector,
  dap_indicator
)
}
\arguments{
\item{validated_analysis}{Analysis file in long format. This file must have stat,key_index,admin_level,admin_name column}

\item{stat}{stat column name in the analysis file}

\item{analysis_key}{Analysis key name in the analysis file}

\item{kobo_survey}{Kobo survey}

\item{kobo_choice}{Kobo choice}

\item{dap}{Data analysis file. Must be the output file from \code{create_dap}.}

\item{dap_sector}{Sector name in the DAP}

\item{dap_indicator}{Indicator name in the DAP}
}
\description{
Title
}
